+++ backend/src/main/java/com/ait/grooming/repository/BreedRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BreedRepository extends JpaRepository<Breed, Long> {
+public interface BreedRepository extends JpaRepository<Breed, Integer> {
 }
