+++ backend/src/main/java/com/ait/grooming/repository/ClientRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 ;
 
 @Repository
-public interface ClientRepository extends JpaRepository<User, Long> {
+public interface ClientRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
 
 }
