+++ backend/src/main/java/com/ait/grooming/repository/GroomingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroomingRepository extends JpaRepository<Grooming, Long> {
+public interface GroomingRepository extends JpaRepository<Grooming, Integer> {
     boolean existsByName(String name);
 }
