+++ backend/src/main/java/com/ait/grooming/repository/ReviewRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReviewRepository extends JpaRepository<Review, Long> {
+public interface ReviewRepository extends JpaRepository<Review, Integer> {
     boolean existsByReview(String review);
 
     boolean existsByAppointment(Appointment appointment);
